consolidated Clock common code and removed unused implementation
